Saving and Loading
Made saving and loading only the first element in the database.

Player will be asked if they want to overwrite their save file.

Player will be prompted their save data when they load if they want to load the data, with name, level, health, monster name, hp and the amount of  inventory items